Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

I got confused by the placement of this new code. It looks like this file belongs to Special:SetLabelDescriptionAliases, but it does not. It belongs to Special:NewItem. The language can be changed there, which changes the placeholders and should also change the directionality. So having this new code as part of this "change" handler is correct.

Thanks for the patch! Just one really minor observation – see my inline comment.

Patch-set: 2
Label: Code-Review=+1