Update patch set 5
Patch Set 5: Code-Review+2

I agree that this is not the nicest way to do this, but definitely better than what we have now.

I discussed the Special:ItemDisambiguation changes with Lydia and we concluded that this is good enough for now. Mid-term we probably want to turn this special page into something different than the search again, or remove it, if it's mostly unused (which seems to be the case now). Anyway, this change is ok for now and helps with the issues at hand.

Patch-set: 5
Reviewer: Hoo man <50@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2