Update patch set 5
Patch Set 5:

I agree that this change is an improvement, and good enough for now. I did not analyze the change in behavior in detail, though.

However, the commit message is bad. It should say "in SpecialItemDisambiguation", not "in several places".

Re @Hoo:
> I discussed the Special:ItemDisambiguation changes with Lydia and we concluded that this is good enough for now. Mid-term we probably want to turn this special page into something different than the search again, or remove it, if it's mostly unused (which seems to be the case now).

Why do you think it's mostly unused? Addshore's initial analysis showed it to be hit about 30 thousand times per day, that's 20 times per minute. Was that analysis incorrect?

Patch-set: 5