Update patch set 1
Patch Set 1:

As said in the other, now abandoned patch: You can as well remove the defaults from all 3 places. Technically they are not needed. They are basically just repeating the documentation of the expected function signatures – which does have some value. So if there is an easy way to keep them, fine. :-)

Patch-set: 1