Update patch set 3
Patch Set 3:

> Ultimately, it is important for Wikibase to embrace the fact that the module loader is not a class loader. We should treat modules similar to how one in other projects might treat Webpack entrypoints for bundle generation. Each module is a deliverable bundle that is registered to the client up front and separately loadable. Unless a module brings in a significant cost, it is usually not worth registering as its own module, in favour of having slightly larger modules, and therefore better caching across views, a much simpler bundle graph, and a much smaller footprint on all page views.

Thanks for pointing this out, Krinkle. This definitely not how Wikibase developers have been dealing with this. I will spread the word then, especially for the future!

Regarding the patch itself: I'd fine with merging it etc. What leaves me a bit uncertain though is the WIP remark in the commit message. Krinkle: would you like this in as it is now, or should I continue from here, and merge then?

Patch-set: 3
Label: Code-Review=0