Update patch set 4
Patch Set 4:

> Wouldn't that prevent the template edit from succeeding as well?

Not necessarily if the description is generated from template parameters, or the name of the page, or other factors which are similar for the template page than the transcluding page.

> Another good reason to not cascade the impact of this parser function, but I fear that horse has bolted. :-(

How would that look in practice? Stop the parser function from taking any effect when it's transcluded? IIRC there isn't any clean way to do that.

Patch-set: 4