Update patch set 1
Patch Set 1:

(1 comment)

As you have probably noticed, mediawiki-core-qunit-selenium-jessie did not run the tests from this repo:

https://integration.wikimedia.org/ci/job/mediawiki-core-qunit-selenium-jessie/20297/console

The reason is that the tests (core and extensions) are executed from core. Configuration file in core expects tests in tests/selenium, but in this repo they are in repo/tests/selenium

https://phabricator.wikimedia.org/source/mediawiki/browse/master/tests/selenium/wdio.conf.js;21ca5a3471dd52c7edb3e13c63fe51134923da6c$47

We could make an exception, like we did for VisualEditor:

https://phabricator.wikimedia.org/source/mediawiki/browse/master/tests/selenium/wdio.conf.js;21ca5a3471dd52c7edb3e13c63fe51134923da6c$48

but since in this repo there already is tests folder, why not put test in tests/selenium?

Patch-set: 1
Reviewer: Zfilipin <487@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0