Update patch set 5
Patch Set 5:

Regarding the structure of files/classes in Wikibase, this is completely understood. It is not "your" code base, in the end.
The comments I made I mostly addressed to my fellow WMDE colleagues to possibly chime in and share their views. Not a problem, we'll take it from here.

Patch-set: 5
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>