Update patch set 1
Patch Set 1:

I would like to see the WMF specific setting as the default, so that if something goes awry and settings are not read, the dumps group is the one that's used. This is how other dump maintenance scripts work, as a safety measure.

Patch-set: 1
Reviewer: ArielGlenn <6@e9e9afe9-4712-486d-8885-f54b72dd1951>