Use `Language::factory(...)` instead of `new Language(...)`
The constructor of Language does not accept an argument and always set
the language code to 'en'. Language objects should created by
Language::factory which ensures that the same language get the same
Language object.

Also change the test from 'qug' to the default 'en' to have the same
behavior as before.

Change-Id: I61439bac82021344c3f9a6056cccd937b3450af2