Update patch set 1
Patch Set 1:

When Language::factory is the right way to get Language objects then this should also used in tests to ensure that caching effects are also tested. Language::newFromCode is not a public method.

The tests fails. This indicates that the tests are wrong. Because 'new Language($code)' is not supported they were never right.

Patch-set: 1
Reviewer: Fomafix <562@e9e9afe9-4712-486d-8885-f54b72dd1951>