LanguageFallbackLabelDescriptionLookupFactory returns buffering lookup
This allows getting rid of the odd optional TermBuffer property,
and especially the side-effect behaviour of creating new lookup
performing some "prefetching".
Now the created lookup takes care of buffering itself.

It seems the only existing use of the factory that didn't
ask for the "buffering/prefetching" lookup has been in
WikibaseRepo::getEntityDiffVisualizerFactory.
It seems that buffering there should not cause harm, even
if the use of the same item/property in diffs is less
likely than in other cases.

Change-Id: Ie0f818f4e0dc7556aed49bec79371a059a1d839a