Stop overriding ContentHandler::makeParserOptions()
A ParserOptions object per content handler doesn't make sense in MCR,
so the ability for ContentHandler to provide one is going to be
deprecated soon. For that to happen, Wikibase needs to stop using it.

To maintain the behavior of splitting the parser cache on
EntityHandler::PARSER_VERSION, we instead register an option on all
ParserOptions objects and trigger it when necessary, as is already done
for 'userlang'.

Bug: T190712
Change-Id: I0712b881ab5391f811352ae3c611652cdd891171