Update patch set 11
Patch Set 11:

> Well, the goal would be to never use php serialization for cache
 > serialization in the first place, this has already caused issues
 > https://phabricator.wikimedia.org/T197252#4292178
 > 
 > > If we are talking about classes like `ItemSerializer`, it is
 > completely different issue
 > We probably will be in the future

Sure! But I would rephrase it as "don't store objects in cache".

I thought about it already.
We can validate stored value and ensure that it doesn't contain any objects, but it requires us to recursively traverse value which might have some performance impact (may be it is minimal and we should not care about it, didn't check). That's why I didn't implement it.

Patch-set: 11