Update patch set 1
Patch Set 1:

Cool.
AFAICT that was the only use of `externalEntityIdParser` in `EntityParserOutputGeneratorFactory` so - even though not really a shortcoming of the current patch - do you not think it would be worthwhile to avoid injecting `EntityIdParser` into the factory as well?

Patch-set: 1
Reviewer: Pablo Grass (WMDE) <5121@e9e9afe9-4712-486d-8885-f54b72dd1951>