Update patch set 1
Patch Set 1:

As already discussed in person - I see this as an attempt to work around a problem which we have not quite understood, so it's a fair attempt. Were it not for the current problem at hand I would however argue that it's a step in the wrong (unintuitive) direction. We should strive for a consistent view and consequently it would be beneficial to have only one spot in the central factory responsible for generating sub views where the message/language provider can be configured, reducing the chance of inadvertently winding up with what we are currently fighting with, a mix of languages in parts of the same page.
tl;dr $textProvider should be a ViewFactory constructor parameter.

Patch-set: 1
Reviewer: Pablo Grass (WMDE) <5121@e9e9afe9-4712-486d-8885-f54b72dd1951>