Update patch set 4
Patch Set 4:

Leszek you are probably right, but I have not looked into this closely.

One thing I remember encountering when writing this is the fact that ParserOpptions::getTargetLanguage() will often return null. The semantics of this is undefined, but IIRC the expectation is that this will result in the site's content language to be used. But for wikibase, that would be wrong, the user language should be used instead.

To avoid this ambiguity and confusion, one would have to ensure that ParserOptions::getTargetLanguage never returns null.

Patch-set: 4
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>