WIP DNM EXP Use parser target language, not user language.
NOTE: this is sufficient to seemingly fix T199983.
But to also fix more subtle problems (e.g. number formatting
in message parameters), all calls to WikibaseRepo::getUserLanguage()
need to be surveyed. 

The alternative is to go back to rendering in the user language on save,
and ensure the correct cache key is used.

Bug: T201170
Change-Id: Iaf2a1d8604e51a36cb85f0ae088731400ac11507