Move state logic from provider to test
The providers are run long before any tests, even before
setUpBeforeClass().  It is therefore possible that global state like
that accessed by MWNamespace::getValidNamespaces() could change for some
reason in between when the provider runs and the test runs.  This
happens somehow with 450433 applied (NamespaceServices service to
replace MWNamespace), so this is a workaround to stop test failures
until the actual problem is hopefully solved at some future date.

Change-Id: I668f8e560ec39e71bda4624216dbdec8cfe42f03