Use PageUpdater & RevisionStore in Wikibase
This change adds some TODOs that can be addressed in followups.
This change tackels the first half of the attached ticket, with the
second half being the ability to configure which slot entities will
be saved to and read from.

This is a re-submit of reverted change Ie499c20da, modified to
address T201194.

Bug: T194729
Change-Id: If94f8cb117c7806f73ee522c42a70c24a290e385