Make entity redirects configurable per entity type
Exctracting EntityRedirectCreationInteractor into an abstract base class
enables other entity type specific redirect services to share the code.
The only difference in redirect creation for different entity types is
in the guard that decides whether the entity is redirectable, which was
extracted into an abstract method.

This limits CreateRedirect and SpecialRedirectEntity to use the
ItemRedirectCreationInteractor, thereby explicitly making them work
only with items. There isn't currently any entity type that gets
excluded by this, so it's not changing existing behavior.

Bug: T200884
Change-Id: I5ec579bd3fc0104f57c713e45472ba25bb24756d