Update patch set 2
Patch Set 2:

> The only real solution is to make MediaInfo and Lexeme part of at least the final +2 jobs.

I note this would make a trick, but there are arguments in favour of not doing this. But this indeed seems to be the "simplest" (as in fastest to set up) way to make it harder to break stuff depending on Wikibase.

>  What is the state of this?

WikibaseLexeme used to be pulled in into Wikibase CI jobs for couple of days until two ago, or so (see e.g. T200693 for starters). This resulted in quite some issues and lead us to, at least temporarily, kick out WikibaseLexeme out.

Patch-set: 2
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>