Update patch set 3
Patch Set 3:

The failure of mwext-Wikibase-repo-tests-sqlite-hhvm-jessie seem to be hinting on the actual issue in Wikibase code (composer does install DM 8.0.0).
Similar regarding the failures in mwext-Wikibase-client-tests-mysql-hhvm-jessie

mwext-php70-phan-docker in turn might be happening as composer.json of WikibaseQualityConstraints says "wikibase/data-model": "~4.2|~5.0|~6.0|~7.0"

and composer fails to compute which version to install.

Patch-set: 3
CC: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>