WIP: Centrally collect various kinds of content languages
This unifies the different kinds of content languages that Wikibase
knows about (getTermsLanguages(), getMonolingualTextLanguages()) in a
single collection and adds a hook to allow extensions to modify that
collection to register their own lists of valid content languages. This
is useful so that we can expose the valid content languages for all
those contexts via a single API module.

(In theory, this also allows extensions to modify the content languages
for Wikibase’ own contexts, e. g. modifying the valid terms languages.)

TODO:
* Decide whether we should alias 'term' to
  'label'/'description'/'alias' or not
* Does this belong in the repo or in the client?

Bug: T198202
Change-Id: If4d6bf8ae9dab25febaf82ad3e04057c8004cfa7