Update patch set 3
Patch Set 3:

> I think we need to create separate builder classes for Form and Sense then, in any case, single class with a switch looks wrong in any case.

Yes, indeed. I was planning to do it in a follow-up change (in order to keep readable diffs).

> But then I wonder why we need a special subEntityMentioned method? It's essentially duplicating entityReferenceMentioned with only difference being that it does not look on RdfProducer::PRODUCE_RESOLVED_ENTITIES flag.

It has also the difference of producing the full description of the entity and not only the "stub". So in your proposal we also need to track this information in the $entitiesResolved data structure of the RdfBuilder class.

To keep things abstract I believe that we want to have the following features:
- allow implementations of EntityRdfBuilder to ask to add the full or the stub serialization of an entity to the output without having to know anything about how they are serialized
- allow RdfProducer to filter the addition of the stub entities
- do not output the stub entity if the full entity is also returned

Patch-set: 3
Reviewer: Tpt <141@e9e9afe9-4712-486d-8885-f54b72dd1951>