Store: add ability to clean in-memory cache
Stores contain an in-memory cache, based on the assumption that within
a single process there is no need for repeated lookups for the same
items - true in production this causes problems in testing.
Static members of WikibaseRepo and its children are not freed between
tests which can lead to tests passing or failing due to state carrying
over.
This creates a member that can be updated (through reflection) when
needed.

Bug: T201988
Change-Id: If9448650a56f30294cc3ef766b5f5ec201e85ba4