Update patch set 10
Patch Set 10:

Interesting approach. No objections from my side.

Not a huge fan of traits and slightly concerned about the tearDown method. Users of the trait need to take care to call it if they have their own extra stuff in a tearDown. I actually don't know how this works for traits. You sure this does not create problems if the test class is already inheriting an important tearDown?

What about using the SomeTestingThing::instance()->getMyStuff() and having SomeTestingThing do the needed cleanup in __destruct()? Idk if that is better though it is probably the approach I would explore first.

Patch-set: 10
CC: Jeroen De Dauw <44@e9e9afe9-4712-486d-8885-f54b72dd1951>