Update patch set 1
Patch Set 1:

Would it be possible to add a comment block in the .owl file stating which extension the constraint stuff is for perhaps with a link to the mw.org page?
Just to help other people that might read this and then be confused about where this constraint stuff happens?

Patch-set: 1