Update patch set 5
Patch Set 5: Code-Review-2

Thank you for the effort of making the patch.

We talked with the colleagues and decided that redirect case is not an exceptional case, so it should be a returned value rather than exception.

I'm currently implementing it in these patches:
  1. https://gerrit.wikimedia.org/r/c/mediawiki/extensions/Wikibase/+/453118
  2. https://gerrit.wikimedia.org/r/c/mediawiki/extensions/Wikibase/+/453148

I'm not sure if result object will fit current code well, but lets see. If it won't we'll use this patch as soon as it is almost ready.

I put `-2` so no-one will accidentally merge this patch while I work on those two.

Any way, creation of this patch is not a wasted effort, because job you've done guides me through the places you changed and helps to understand where and how I need to change the code.

Sorry for how things worked out.

Patch-set: 5
Reviewer: Aleksey Bekh-Ivanov (WMDE) <4089@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2