Update patch set 5
Patch Set 5:

(1 comment)

To clear up "Looks pretty good." & the -1

This patch looks pretty good, and I'm essentially ready to +2 and merge it once the 2 comments that require changes are fixed and we have made sure that this isn't going to break any extensions as outlines in one of my comments.

If the patch owner doesn't fix the comments in the next few days then I probably would and +2 it.

However now I hear that we want to investigate a different route, so I'll hold off:
 - https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/Wikibase/+/453118/
 - https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/Wikibase/+/453148/
 - https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/WikibaseLexeme/+/453149/

Patch-set: 5
Reviewer: Addshore <106@e9e9afe9-4712-486d-8885-f54b72dd1951>