Update patch set 4
Patch Set 4:

One of the intentions was to force a caller to handle all the cases. I came up with this requirement because I've recently introduced a bug in a new code just because I didn't think of the fact that revision can be a redirect (which IMO is not obvious) and should be treated differently in some parts of the code. For the code that doesn't care if it is a redirect or it is just a revision, it is pretty trivial to handle both cases in a same way ($revisionId is the first argument in both handlers).
Also, if a case is added in the future, all the unchanged code will instantly fail requiring explicit handling.

Having a handler as a second argument means you cannot pass result as a value for other code to handle. Also, in your example, `getLatestRevisionId` returns `void`, which means caller will have to use references in handler to work with. To avoid it, the correct way in PHP will be for `getLatestRevisionId` to return `mixed` and throw `\Exception`.

Regarding the caller version: I created 3 different methods on the result because having a constructor with 3 `callable` arguments means that it is quiet easy to mess up the order. Also, it might be not that clear for the reader to see which closure handles which case.

To be fair, I also don't like a necessity to call `map()` to get the value, but couldn't find any other way.

I got the idea from pattern matching in Scala, Rust and Haskell, where compiler tells you that you "switch-case" is not exhaustive and you forgot to handle a case. 

Comparing `getLatestRevisionId` to your example with UseCases: there is a difference in how the code is used. In case of a UseCase, result, probably, won't be processed by any code (except framework and some middle-ware), but here we need these values to use them in other code.

I didn't finish working with the patch and not sure what the calling code will look like (may be it will be a horrible mess and I'll ditch the idea then)

> Patch Set 4:
> 
> This seems like a pretty unusual approach. Where/how did you get this idea?
> 
> Note that linked code below is pseudo-code-ish with quickly picked names that might not be consistent.
> 
> If there is a result object, I would expect something more along the lines of https://gist.github.com/JeroenDeDauw/5f096852b322233931bdd6bd2f180ffd. That does not force the caller to deal with all cases, it just forces them to make sure they can deal with the cases relevant to them. I'm not sure that is better though my first impression is that it would be.
> 
> If you go with the callback approach then I don't see why a result object is needed. The callbacks look like events, but the info is already there, which is what makes the approach seem weird to me. And then you need to call map() for the thing to figure out which not-event to fire. Instead you can pass the callbacks to the function and skip having the result object or the need for a special invocation trigger (->map()). https://gist.github.com/JeroenDeDauw/b0558444a107d642ee99c40bc2cc9daf
> 
> The latter approach is used in The Clean Architecture by injecting presenters into UseCases with multiple "result" types. These presenters then have methods such as showNotFound(), showAccessForbidden() and showLolcats().

Patch-set: 4
Reviewer: Aleksey Bekh-Ivanov (WMDE) <4089@e9e9afe9-4712-486d-8885-f54b72dd1951>