Update patch set 4
Patch Set 4:

(1 comment)

I am a bit concerned by planning to use EntityMentionListener::entityReferenceMentioned for including forms and senses : forms and senses are parts of the lexemes and so they have a stronger relation than just usual reference to other entities. This distinction make some part of code ackwards like not caring about RdfProducer::PRODUCE_RESOLVED_ENTITIES when the full entity is requested.

A call to RdfBuilder::resolveMentionedEntities should probably be added to the RdfDumpGenerator class in order to add form and senses to the dump.

If we use this design we should also check that we do not do additional not useful SQL query or deserialization for each entity.

Patch-set: 4
Reviewer: Tpt <141@e9e9afe9-4712-486d-8885-f54b72dd1951>