Update patch set 4
Patch Set 4:

>> If we use this design we should also check that we do not do additional not useful SQL query or deserialization for each entity.
> Not sure which queries do you mean, could you explain? It'd be still treating forms/senses as separate entities, so it'd have to load/deserialize them, but that was for previous way too, not?

Sorry, I expressed myself badly. In your implementation, only the form or sense id is going to be stored by RdfBuilder to be serialized later. So, an extra call to EntityLookup is going to be done by the RdfBuilder to retrieve the Form or Sense object from its id. This lookup may lead to extra SQL queries and JSON deserialization if the Lexeme object that contains the Form/Sense is not in the EntityLookup cache.

Patch-set: 4
Reviewer: Tpt <141@e9e9afe9-4712-486d-8885-f54b72dd1951>