Update patch set 6
Patch Set 6:

(8 comments)

Cool, looks generally nice and useful.

One thing that I'm not a big fan of is that the behavior of this entity selector starts to become context specific (e.g. selector for statement value vs outside), almost as if we should have different kinds of them. Also, passing the property id along to the hook is sketchy and seems to be specific to a certain use case for this hook you have in mind. Again, this is a sign that in this case we may want to have an entity selector that knows which property id it's looking up values for.

Patch-set: 6
Reviewer: Jakob <1786@e9e9afe9-4712-486d-8885-f54b72dd1951>