Update patch set 4
Patch Set 4:

> I am still not happy that entitiesResolved in now used in a way it is, but that may not be relevant for this patch.

Yes, definitely. We should maybe refactor and rename EntityMentionListener to something more general in order to make it the interface used to ask for the serialization of any other entity. But it may be done in a follow up change.

Patch-set: 4
Reviewer: Tpt <141@e9e9afe9-4712-486d-8885-f54b72dd1951>