Update patch set 9
Patch Set 9:

(1 comment)

It looks I was having ideas how to possible re-arrange this result object a bit, that were similar to what Jeroen was mentioning.
Having the "result handler" interface was introducing some other issues, so didn't seem a perfect solution either.

I think the current approach in the patch is good enough to be used for now. It can be of course changed once something more "clean" is suggested. Maybe it would make sense that you guys, Aleksey and Jeroen, pair up on trying to do some improvements to it.

Patch-set: 9
CC: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>