Introduce EntityDocumentView interface
This is an alternative to Ia69220d91 & I828e29103e
which change the visibility of various EntityView properties
for use by WikibaseMediaInfo.

This new interface could also be used by WikibaseLexeme.

Bug tagged as the realtes WikibaseMediaInfo ticket.

In the future it would be good to renmove the labels/description
/aliases specific stuff from EntityView and perhaps kill EntityView
all together.

The Lexeme Experience team were recently looking at getTitle from
EntityView so it would be good to have them review this ticket and
add some thoughts.

Bug: T202287
Change-Id: I84dd340aafc3dffd988c1456bf0d80995cd3cb65