Update patch set 6
Patch Set 6: Code-Review-1

(2 comments)

> I can then derive my own OsmItemId from the ItemId class […]

This patch does have pretty much the exact same problem as we discussed in Ia69220d just a few days ago.

Overriding the contract of a class as essential as ItemId and messing with the guarantees it originally gave is a dark pattern and should not be done. The only reason none of the classes in the Wikibase codebases is marked as "final" is because that would be a burden and clutter the code quite a lot. But users of the code should use these classes as if they are "final".

I suggest to read up on LSP: https://en.wikipedia.org/wiki/Liskov_substitution_principle

Patch-set: 6
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1