Remove @embed from wikibase.client.css background-image
The image is usually outside the viewport and should not delay
the rendering of page layout and text through the blocking download
of this icon.

Removing the @embed from the idle icon (grayscale) as should
is fairly straight-forward.

The slightly tricker part is the "hover" state icon. One of the
benefit of @embed is that the icon is downloaded even if the
CSS rule is not currently used. While bad for performance, it did
mean that when hovering the link, the alternate state is immediately
available. Normally in web browsers there is undesirrable behaviour
here where the icon disappears on-hover, and the hover state is
not shown until the image is downloaded from the network, which
can easily take half a second.

In other web apps, this is mitigated by pre-loading the image
in JavaScript after the page is ready. But in ResourceLoader
we always used @embed to avoid this problem. Removing it means
we have to solve it a different way. The upcoming solution for
this is a new annotation @push, but that's not ready yet (T120984).

Instead, this particular use case in Wikibase seems like a case
where really we don't need a separate image, because built-in
CSS is already capable of "dimming" the image. Specifically,
the 'opacity' and 'filter' rules.

For a visual comparison, see
<https://phabricator.wikimedia.org/T121730#4547352>.

Bug: T121730
Change-Id: I539f19c402923865f912f415044b95d2b24ac644