Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review-1
> 
> Wikibase's phan config is set up to use stubs of extensions rather than analyzing the actual extensions themselves. So you'll need to add this class to tests/phan/stubs/cirrussearch.php.

OK, thanks for the note. Should this be documented somewhere and/or eventually fixed? It's not out of the question for extensions to use other extension's classes and having stub setups for each of them just for the benefit of phan does not seem like a good long-term habit.

Patch-set: 4
Reviewer: Smalyshev <2096@e9e9afe9-4712-486d-8885-f54b72dd1951>