EntityReferenceExtractor, no guarantee that array elements will be unique
In Ie56adf738cb I introduced a new implementation of
EntityReferenceExtractor.

While looing at the current implementations of the interface I found
that there was no consistent decision in regards to returning a unique
array or leaving that to the caller.
As a result I decided to add this comment to the interface clarifying
that there is no gaurentee of uniquness.

As a followup it would be nice to go through the implementations and
remove the array_uniquee calls and move them to the callers if actually
needed.

Change-Id: I8c80cbe3c68551acc65474186bb0ab7548336227