Correctly append browser test language param
The browser tests currently only work if the language query parameter
is the only one in the URL. URLs e.g. as generated by the popular docker
solution for mediawiki development look like this:
http://default.web.mw.localhost:8080/mediawiki/index.php?title=Item:Q1

Without this patch, the cucumber tests will fail for such URLs because
it will simpy append `?setlang=en` when it should be `&setlang=en`.

Change-Id: I1f806837610a103ff48952e5004968327f153145