Missing favicon.ico is not a Selenium error
After each scenario, the browser logs are checked and any SEVERE error
will cause the scenario to fail. When the target wiki does not have a
favicon.ico, that triggers a 404 which causes all scenarii to fail.

Filter out the favicon.ico 404 message from the browser logs.

We have done the same patch for WikibaseLexeme:
https://gerrit.wikimedia.org/r/#/c/457427/

CI will be switched from Firefox to Chrome with the next change:
https://gerrit.wikimedia.org/r/#/c/460511/

Bug: T203178
Change-Id: I4177c3312ad8f5a53f23e22aca51aa6bb507977e