SqlStore wrap PropertyInfo classes in an APC cache too
This patch wraps the already cached property info table in another
cache.

Now apc will be checked (with a short ttl), followed by memcached
(longer ttl) and finally hiting the DB itself.

In the future we might want to make the TTL of the apc based cache
configurable but I don't think that is needed right now.
We also might want to tweak it once this is deployed and we see how it
performs.

Bug: T97368
Change-Id: I932f55382bf34e9e697446078558f1adf890b317