Update patch set 2
Patch Set 2:

I would like to understand what issue this change is actually solving? Can you please describe this in the commit message?

As far as I can see all this patch does is making the code run a little faster, because the additional validation is just not necessary. But that's about it. What user-facing issue can be cause by this code? Every newly generated GUID *must* be valid. If it is not, that would be a major issue somewhere else and should not be worked around by disabling the validation.

Patch-set: 2
CC: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>