Adjust wb_terms SQL indexes to match WMF production
This adds 'tmp1' index to the Wikibase install SQL files

At some point in 2018 we thought the index "tmp1"
that exists in WMF production for wikidata
on the wb_terms table was not used.
While removing the index we had a rather large outage,
probably to do with some usage of the index,
and thus the removal was reverted.
In order to being our code SQL files in line
with WMF production this index was added the code.

This also changes term_text, term_search_key and wb_terms_entity_id indices

Detail information about the structure before
the patch see: https://phabricator.wikimedia.org/T85414#4597677

Also for some compatibility reasons
the old indices (term_text, term_search_key) are still in the sql
files and will run by the DatabaseSchemaUpdater.

Bug: T202265
Bug: T204837
Bug: T204838
Bug: T204836
Change-Id: I152ca2233afbdc7a9c95b27eb5e649f61e901ede