Add 'tmp1' index to the Wikibase install SQL files
At some point in 2018 we thoughts the index "tmp1" that exists in WMF production for wikidata on the wb_terms table was not used.
While removing the index we had a rather large outage, probably to do with some usage of the index, and thus the removal was reverted.
In order to being our code SQL files in line with WMF production this index was added the code.

Change-Id: Iac7987f9491ec4b4eabfbde9e557b8731e17b644
Task: T202265