Improve the performance of EntityContent::collectValues
While testing this against the data of Q56596767 on wikidata.org
this change improves the time spent on the getTextForFilters call
that calls collectValues from around 0.8s to 0.015s.
The improvement holds true for php71 and also hhvm.

This is a dramatic performance increase.
getTextForFilters is called twice on edits, for both the old and
new entity.

Bug: T204109
Change-Id: I974f8f36608d350427c83d4e4f1213bbead69816