Track entity save times & sizes in detail in statsd
T204109 reveals some possible speed improvements to be
made in the way Wikibase integrates with AbuseFilter.
In order to track any improvements we first have to track
where we are now.
We track both the speed of saves and the sizes of the
entities being saved.
The bigger the entity the longer the save time to be
expected..

Per suggestion this patch uses decorators for EditEntity
and other services that we want to time.
This patch also moves EditEntity and related classes into
their own namespace.

TODO:
 - Test the Timed classes, possibly using wmde/clock
   once released..
 - Factor out stat collection of entity size at save time into
   some sort of decorator?
 - Figure out if submiting the stats before the entity save has finished
   is a good or bad thing? It could result in tracking more filter run
   times than actual save times for example. We only want to submit on
   success

Bug: T204109
Change-Id: I8cda1ad96ca7dc04924ddf7712fefabedfd7cbd7