Track entity save times & sizes in detail in statsd
T204109 reveals some possible speed improvements to be
made in the way Wikibase integrates with AbuseFilter.
In order to track any improvements we first have to track
where we are now.
We track both the speed of saves and the sizes of the
entities being saved.
The bigger the entity the longer the save time to be
expected..

Bug: T204109
Change-Id: I8cda1ad96ca7dc04924ddf7712fefabedfd7cbd7