Update patch set 2
Patch Set 2:

So, as Thiemo already suspected, the plan was to sneakily work around GUID validation for statements on new Forms that do not yet have an id. This was only planned as an intermediate step until we find a better way (e.g. get rid of entity id prefixes on GUIDs), but we decided to go a different, hopefully less hacky route.

Patch-set: 2
Reviewer: Jakob <1786@e9e9afe9-4712-486d-8885-f54b72dd1951>