DNM: StatsdDataFactoryInterface alternative spike
This gets rid of the low level details where you want to profile.
It also hides the StatsdDataFactoryInterface and is much easier to
have an alternative implementation for. Decreases the likelyhood
that decorators are warranted to do this kind of profiling.

Downsides:
* Return value is passed via recordTiming(). Not the most obvious
* (minor) Callback needs parent scope arguments which is awkward in PHP

I did not find a way to avoid these downsides without introducing bigger
ones.

Change-Id: I446c17050ff0688f10c985786b595b1344291849