Fix keyboard navigation when editing statements
With this change, tabbing to links like “add qualifier”, “add reference”
or “remove” while editing a statement and then pressing enter will no
longer save the entire statement. This has been broken for at least a
year, and meant that it was effectively impossible to add statements
with qualifiers or references only using the keyboard, since the first
attempt to start adding qualifiers or references would already save the
entire statement. This finally makes that possible again.

Bug: T154869
Change-Id: I6e815c91e54a1ace77ac52da4385c6ad0fe03c3d