Update patch set 2
Patch Set 2:

> > Patch Set 2:
 > >  > Could be I'm misunderstanding something here, of course.  If
 > this
 > >  > method were used in more than one place I'd be tempted to
 > refactor
 > >  > it to make it more clear whether we're talking about page
 > language
 > >  > or what...
 > >
 > > Ah, maybe I'm misunderstanding.  This use site is passing the
 > "context" of the article, ie UI context, not a context representing
 > the "content" of the article?
 > 
 > Wikibase doesn't have a content language. All content is always
 > rendered in the UI language. The data itself is language neutral.

Ok.  In the context of this particular bug, you've got content pre-rendered in zh-hans/zh-hant/etc (that is, messages localized directly into the variant) and so don't want/need LanguageConverter to run on the output.  Just double-checking that this is the case...

Patch-set: 2
Reviewer: C. Scott Ananian <664@e9e9afe9-4712-486d-8885-f54b72dd1951>