Update patch set 3
Patch Set 3:

> Patch Set 2:
> Ok.  In the context of this particular bug, you've got content pre-rendered in zh-hans/zh-hant/etc (that is, messages localized directly into the variant) and so don't want/need LanguageConverter to run on the output.  Just double-checking that this is the case...

This is the case for the actual content, but I'm not 100% sure this is the case for the system messages used when generating the rendered entity. It's theoretically possible that some of these messages rely on variant conversion, I have not checked.

Patch-set: 3
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>