Update patch set 11
Patch Set 11: Code-Review+1

Would be nice if we could have a test case that ensures that WikiPageEntityRevisionLookup behaves correctly when the page exists but the slot does not. WikiPageEntityRevisionLookupTest is an integration test that hits the database, so it would fit in there. But that would require the store to be set up to use slots.

+2 otherwise. The new tests can also go into a follow-up.

Patch-set: 11
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1