Update patch set 4
Patch Set 4: Code-Review-1

One issue with this and the implementation of the interface in media info is that indexResultByEntityId will no longer work as that indexes the results by page_title too.

We need to index by entity id, so possible we should just add that to the select?

Patch-set: 4
Reviewer: Addshore <106@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1