Update patch set 3
Patch Set 3:

Yes, unfortunately I believe this is going to be a long-running side project. I try to tend to it every now and then but its not a small feat and the performance implications are maybe too big to just go for it + the solution is not almost obvious + local caches (e.g. static members) are not easily removed without much reverse engineering.

Patch-set: 3
Reviewer: Pablo Grass (WMDE) <5121@e9e9afe9-4712-486d-8885-f54b72dd1951>